requirements.txt
python-docx
docxtpl
PyPDF2==3.0.1
fpdf2>=2.7.0
Flask==3.0.0
customtkinter
Pillow
pyperclip
requests
opencv-python
pytesseract
numpy
pdfkit
pdf2image
python-dotenv
google-generativeai
PyQt6
PyQt6-WebEngine

